+++ docs/source/release_notes.rst
@@ -16,6 +16,7 @@ Release Notes
         * Remove WholeNumber LogicalType from Woodwork (:pr:`380`)
         * Updates to LogicalTypes to support Koalas 1.4.0 (:pr:`393`)
         * Replace ``set_logical_types`` and ``set_semantic_tags`` with just ``set_types`` (:pr:`379`)
+        * Remove ``copy_dataframe`` parameter from DataTable initialization (:pr:`398`)
     * Documentation Changes
     * Testing Changes
         * Add pyarrow, dask, and koalas to automated dependency checks (:pr:`388`)
@@ -29,6 +30,7 @@ Release Notes
     * The ``DataTable.set_logical_types`` method was removed. ``DataTable.set_types`` can be used instead.
     * ``WholeNumber`` was removed from LogicalTypes. Columns that were previously inferred as WholeNumber will now be inferred as Integer.
     * The ``DataTable.get_mutual_information`` was renamed to ``DataTable.mutual_information``.
+    * The ``copy_dataframe`` parameter was removed from DataTable initialization.
 
 **v0.0.5** November 11, 2020
     * Enhancements